Separated code out of index.js so it isn't as massive
Better organized imports. Aswell took neo4j logic and building the goldenLayout out of index.js.
